pyproject.toml
[tool.poetry]
name = "fastapi-idade-sql"
version = "0.1.0"
description = ""
authors = ["LivMad <livia.o.dantas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlmodel = "^0.0.18"
ulid-py = "^1.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
